script only saves every 10 epochs

+++ scripts/segment_classifier_model_train.py
@@ -156,11 +156,12 @@ if __name__ == '__main__':
 
         log_loss_to_csv(epoch,avg_train_loss,avg_val_loss, avg_pixel_acc_loss, avg_dice_loss, avg_iou_loss, save_location)
         
-        # Remove previous model if it exists
-        if epoch > 0:
-            prev_model_path = f'{save_location}model_{epoch}.pth'
-            if os.path.exists(prev_model_path):
-                os.remove(prev_model_path)
-        
-        # Save current model
-        torch.save(model.state_dict(), f'{save_location}model_{epoch+1}.pth')
+        if epoch % 10 == 0:
+            # Remove previous model if it exists
+            if epoch > 0:
+                prev_model_path = f'{save_location}model_{epoch}.pth'
+                if os.path.exists(prev_model_path):
+                    os.remove(prev_model_path)
+            
+            # Save current model
+            torch.save(model.state_dict(), f'{save_location}model_{epoch+1}.pth')
